Review comments and renames
