+++ api/src/main/java/org/apache/iceberg/actions/ActionsProvider.java
@@ -57,7 +57,7 @@ public interface ActionsProvider {
   /**
    * Instantiates an action to rewrite data files.
    */
-  default RewriteDataFiles rewriteDataFiles(Table table) {
+  default RewriteDataFilesV1 rewriteDataFiles(Table table) {
     throw new UnsupportedOperationException(this.getClass().getName() + " does not implement rewriteDataFiles");
   }
